refactor: give the path rewriting its own class and its own tests
RequestPathCanonicalizer holds what the filter used to carry as private statics: percent-decode,
collapse duplicate separators, remove dot segments, in the container's order. It is a bundle
class, not part of the exported API, so it stays an implementation detail of this module rather
than a URL utility other bundles would depend on a cache-control API for.

RequestPathCanonicalizerTest covers the rewriting on its own, and ClientCacheFilterTest is left
with the question that belongs to the filter: which path it hands to the service.
